# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Craftedby API",
        "description": "The Craftedby API provides comprehensive endpoints to manage products and users. It supports CRUD operations for products. User management features include user registration, authentication, and profile management. Additionally, the API supports product search and category filtering. Designed for scalability and performance, it ensures robust and secure interactions between clients and the eCommerce backend.",
        "contact": {
            "email": "nathalie.soto@le-campus-numerique.fr"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a list of all products",
                "description": "Display a listing of the resource.",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "description": "Store a newly created resource in storage.",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "story",
                                    "price",
                                    "stock",
                                    "image",
                                    "category"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Product name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "story": {
                                        "type": "string",
                                        "example": "Product story"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "Product image"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Product category"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Product color"
                                    },
                                    "material": {
                                        "type": "string",
                                        "example": "Product material"
                                    },
                                    "size": {
                                        "type": "string",
                                        "example": "Product size"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product stored"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product by ID",
                "description": "Display the specified resource.",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "401": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to authenticate",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}